requirements.txt
mesa>=2.0.0
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
pyyaml>=6.0
